Extract permission handling in separate library
