+++ common-libraries/common-permission-handling/src/main/java/by/vsu/uis/common/permission/PermissionHandlingConfig.java
@@ -1,0 +1,9 @@
+package by.vsu.uis.common.permission;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan("by.vsu.uis.common.permission")
+public class PermissionHandlingConfig {
+}
