+++ common-libraries/common-permission-handling/src/main/java/by/vsu/uis/common/permission/PermissionProducer.java
@@ -1,4 +1,4 @@
-package by.kovzov.uis.security.service.extract;
+package by.vsu.uis.common.permission;
 
 import java.util.List;
