+++ common-libraries/common-permission-handling/src/main/java/by/vsu/uis/common/permission/PreAuthorizePermissionProducer.java
@@ -1,4 +1,4 @@
-package by.kovzov.uis.security.service.extract;
+package by.vsu.uis.common.permission;
 
 import java.util.List;
 import java.util.Map;
