+++ common-libraries/pom.xml
@@ -7,6 +7,7 @@
         <module>common-exception-handling</module>
         <module>common-exceptions</module>
         <module>common-validator</module>
+        <module>common-permission-handling</module>
     </modules>
     <parent>
         <groupId>org.springframework.boot</groupId>
