+++ security-server/security-server-rest/src/main/java/by/kovzov/uis/security/rest/SecurityServerApplication.java
@@ -1,12 +1,14 @@
 package by.kovzov.uis.security.rest;
 
 import by.kovzov.uis.security.service.api.StartUpService;
+import by.vsu.uis.common.permission.PermissionHandlingConfig;
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Import;
 
 @SpringBootApplication(scanBasePackages = {
     "by.kovzov.uis.security",
@@ -16,6 +18,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @OpenAPIDefinition
 @AllArgsConstructor
 @Slf4j
+@Import(value = {
+    PermissionHandlingConfig.class
+})
 public class SecurityServerApplication implements CommandLineRunner {
 
     private final StartUpService startUpService;
