+++ security-server/security-server-service/pom.xml
@@ -23,6 +23,11 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
+            <groupId>by.kovzov.uis.common</groupId>
+            <artifactId>common-permission-handling</artifactId>
+            <version>0.0.1-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>by.kovzov.uis.security</groupId>
             <artifactId>security-server-dto</artifactId>
             <version>${project.version}</version>
@@ -33,10 +38,6 @@
             <version>0.0.1-SNAPSHOT</version>
         </dependency>
 
-      <!--  <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-security</artifactId>
-        </dependency>-->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-oauth2-resource-server</artifactId>
